+++ fearless_simd_gen/src/mk_simd_trait.rs
@@ -155,7 +155,7 @@ fn mk_simd_base() -> TokenStream {
         pub trait SimdBase<S: Simd>:
             Copy + Sync + Send + 'static
             + Seal
-            + Bytes + SimdFrom<Self::Element, S>
+            + Bytes + SimdFrom<Self::Element, S> + SimdFrom<Self::Array, S>
             + core::ops::Index<usize, Output = Self::Element> + core::ops::IndexMut<usize, Output = Self::Element>
             + core::ops::Deref<Target = Self::Array>+ core::ops::DerefMut<Target = Self::Array>
         {
@@ -176,7 +176,12 @@ fn mk_simd_base() -> TokenStream {
             /// The array type that this vector type corresponds to, which will
             /// always be `[Self::Element; Self::N]`. It has the same layout as
             /// this vector type, but likely has a lower alignment.
-            type Array;
+            type Array: Copy
+                + core::fmt::Debug
+                + IntoIterator<Item = Self::Element>
+                + AsRef<[Self::Element]>
+                + AsMut<[Self::Element]>
+                + From<Self>;
             /// Get the [`Simd`] implementation associated with this type.
             fn witness(&self) -> S;
             fn as_slice(&self) -> &[Self::Element];